requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
alembic==1.11.1
pydantic==1.10.7
bcrypt==4.3.0
